CRAB-34668, CRAB-32920, CRAB-33972: Add-on Packaging Utility (#6)
This is the very first version of the Add-on Packaging Utility. The Add-on Packaging Utility is a CLI (Command Line Interface) utility that uses a templated project in the backend to walk the user through a series of prompts. When the user answers all the prompts, the utility creates a new project with working examples of packaged Add-ons that can be readily deployed to Seeq via the Add-on Manager.